requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
distro==1.8.0
Django==4.2.4
django-cachalot==2.6.1
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-filter==23.2
django-js-asset==2.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.5
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
kombu==5.3.1
MarkupSafe==2.1.3
packaging==23.1
paramiko==3.3.1
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
pytz==2023.3
PyYAML==3.10
redis==5.0.0
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
texttable==1.6.7
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
websocket-client==0.59.0
